+++ src/main/java/kitchenpos/order/delivery/domain/model/DeliveryOrder.java
@@ -1,0 +1,56 @@
+package kitchenpos.order.delivery.domain.model;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.ForeignKey;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToMany;
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.UUID;
+import kitchenpos.order.order.model.OrderLineItem;
+import kitchenpos.order.order.model.OrderStatus;
+
+@Entity
+public class DeliveryOrder {
+    @Id
+    @GeneratedValue
+    private UUID id;
+
+    @Enumerated
+    private DeliveryOrderFlow orderFlow;
+
+    @Column(name = "order_date_time", nullable = false)
+    private LocalDateTime orderDateTime;
+
+    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
+    @JoinColumn(
+            name = "order_id",
+            nullable = false,
+            columnDefinition = "binary(16)",
+            foreignKey = @ForeignKey(name = "fk_order_line_item_to_orders")
+    )
+    private List<OrderLineItem> orderLineItems;
+
+    @Column(name = "delivery_address")
+    private String deliveryAddress;
+
+    public boolean validateOrderFlowAndFindNextStep(OrderStatus orderStatus) {
+        if (orderFlow.validateOrderStatus(orderStatus)) {
+            throw new IllegalArgumentException();
+        }
+        return orderFlow.isRiderNecessary(orderStatus);
+    }
+
+    public String getDeliveryAddress() {
+        return this.deliveryAddress;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+}
