+++ src/main/java/kitchenpos/order/delivery/domain/port/RiderPort.java
@@ -1,0 +1,10 @@
+package kitchenpos.order.delivery.domain.port;
+
+import kitchenpos.order.delivery.domain.model.DeliveryOrder;
+import org.springframework.stereotype.Component;
+
+@Component
+public interface RiderPort {
+
+    void requestRider(DeliveryOrder order, String deliveryAddress);
+}
