+++ src/main/java/kitchenpos/order/delivery/domain/repository/DeliveryRepository.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.delivery.domain.repository;
+
+public interface DeliveryRepository {
+}
