+++ src/main/java/kitchenpos/order/delivery/infra/external/KitchenridersClient.java
@@ -1,0 +1,8 @@
+package kitchenpos.order.delivery.infra.external;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public interface KitchenridersClient {
+    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
+}
