+++ src/main/java/kitchenpos/order/delivery/infra/persistence/JpaDeliveryOrderRepository.java
@@ -1,0 +1,6 @@
+package kitchenpos.order.delivery.infra.persistence;
+
+import kitchenpos.order.delivery.domain.repository.DeliveryRepository;
+
+public interface JpaDeliveryOrderRepository extends DeliveryRepository {
+}
