+++ src/main/java/kitchenpos/order/delivery/service/DeliveryOrderService.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.delivery.service;
+
+public class DeliveryOrderService {
+}
