+++ src/main/java/kitchenpos/order/delivery/ui/DeliveryOrderController.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.delivery.ui;
+
+public class DeliveryOrderController {
+}
