--- src/main/java/kitchenpos/order/domain/eatin/flow/EatInOrderFlow.java
@@ -1,4 +1,0 @@
-package kitchenpos.order.domain.eatin.flow;
-
-public class EatInOrderFlow {
-}
