--- src/main/java/kitchenpos/order/domain/eatin/model/EatInOrder.java
@@ -1,4 +1,0 @@
-package kitchenpos.order.domain.eatin.model;
-
-public class EatInOrder {
-}
