--- src/main/java/kitchenpos/order/domain/order/model/OrderStatus.java
@@ -1,5 +1,0 @@
-package kitchenpos.order.domain.order.model;
-
-public enum OrderStatus {
-    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
-}
