--- src/main/java/kitchenpos/order/domain/order/model/OrderType.java
@@ -1,5 +1,0 @@
-package kitchenpos.order.domain.order.model;
-
-public enum OrderType {
-    DELIVERY, TAKEOUT, EAT_IN
-}
