--- src/main/java/kitchenpos/order/domain/order/repository/OrderTableRepository.java
@@ -1,15 +1,0 @@
-package kitchenpos.order.domain.order.repository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-import kitchenpos.order.domain.eatin.model.OrderTable;
-
-public interface OrderTableRepository {
-    OrderTable save(OrderTable orderTable);
-
-    Optional<OrderTable> findById(UUID id);
-
-    List<OrderTable> findAll();
-}
-
