--- src/main/java/kitchenpos/order/domain/takeout/model/TakeOutOrder.java
@@ -1,4 +1,0 @@
-package kitchenpos.order.domain.takeout.model;
-
-public class TakeOutOrder {
-}
