+++ src/main/java/kitchenpos/order/eatin/model/EatInOrder.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.eatin.model;
+
+public class EatInOrder {
+}
