+++ src/main/java/kitchenpos/order/eatin/ui/EatInOrderController.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.eatin.ui;
+
+public class EatInOrderController {
+}
