--- src/main/java/kitchenpos/order/infrastructure/external/KitchenridersClient.java
@@ -1,8 +1,0 @@
-package kitchenpos.order.infrastructure.external;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-public interface KitchenridersClient {
-    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
-}
