+++ src/main/java/kitchenpos/order/takeout/model/TakeOutOrder.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.takeout.model;
+
+public class TakeOutOrder {
+}
