+++ src/main/java/kitchenpos/order/takeout/model/TakeOutOrderFlow.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.takeout.model;
+
+public class TakeOutOrderFlow {
+}
