+++ src/main/java/kitchenpos/order/takeout/ui/TakeOutOrderController.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.takeout.ui;
+
+public class TakeOutOrderController {
+}
