--- src/main/java/kitchenpos/order/ui/OrderTableRestController.java
@@ -1,56 +1,0 @@
-package kitchenpos.order.ui;
-
-import kitchenpos.order.domain.eatin.application.OrderTableService;
-import kitchenpos.order.domain.eatin.model.OrderTable;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.net.URI;
-import java.util.List;
-import java.util.UUID;
-
-@RequestMapping("/api/order-tables")
-@RestController
-public class OrderTableRestController {
-    private final OrderTableService orderTableService;
-
-    public OrderTableRestController(final OrderTableService orderTableService) {
-        this.orderTableService = orderTableService;
-    }
-
-    @PostMapping
-    public ResponseEntity<OrderTable> create(@RequestBody final OrderTable request) {
-        final OrderTable response = orderTableService.create(request);
-        return ResponseEntity.created(URI.create("/api/order-tables/" + response.getId()))
-            .body(response);
-    }
-
-    @PutMapping("/{orderTableId}/sit")
-    public ResponseEntity<OrderTable> sit(@PathVariable final UUID orderTableId) {
-        return ResponseEntity.ok(orderTableService.sit(orderTableId));
-    }
-
-    @PutMapping("/{orderTableId}/clear")
-    public ResponseEntity<OrderTable> clear(@PathVariable final UUID orderTableId) {
-        return ResponseEntity.ok(orderTableService.clear(orderTableId));
-    }
-
-    @PutMapping("/{orderTableId}/number-of-guests")
-    public ResponseEntity<OrderTable> changeNumberOfGuests(
-        @PathVariable final UUID orderTableId,
-        @RequestBody final OrderTable request
-    ) {
-        return ResponseEntity.ok(orderTableService.changeNumberOfGuests(orderTableId, request));
-    }
-
-    @GetMapping
-    public ResponseEntity<List<OrderTable>> findAll() {
-        return ResponseEntity.ok(orderTableService.findAll());
-    }
-}
