+++ src/test/java/kitchenpos/order/order/infrastructure/persistence/InMemoryOrderRepository.java
@@ -1,9 +1,9 @@
 package kitchenpos.order.order.infrastructure.persistence;
 
-import kitchenpos.order.domain.order.model.Order;
-import kitchenpos.order.domain.order.repository.OrderRepository;
-import kitchenpos.order.domain.order.model.OrderStatus;
-import kitchenpos.order.domain.eatin.model.OrderTable;
+import kitchenpos.order.order.model.Order;
+import kitchenpos.order.order.repository.OrderRepository;
+import kitchenpos.order.order.model.OrderStatus;
+import kitchenpos.order.eatin.model.OrderTable;
 
 import java.util.ArrayList;
 import java.util.HashMap;
